Refactor ROIL-P to estimate the expert's policy and prune reward functions
This commit refactors the ROIL-P algorithm in the main.tex file. The key strength of ROIL-P is that it does not estimate the expert's occupancy frequency, but instead estimates the expert's policy. In on-policy domains, estimates of the expert's policy are close to the true expert. The commit also prunes the set of reward functions based on the estimated policy. The code changes include modifying the text and equations in the ROIL-P section.